// src/types/sekolah.ts
export interface ProfilSekolah {
  id: number;
  header_baris1: string | null;
  header_baris2: string | null;
  nama_sekolah: string | null;
  nss: string | null;
  npsn: string | null;
  alamat: string | null;
  kelurahan: string | null;
  kecamatan: string | null;
  kota_kabupaten: string | null;
  kode_pos: string | null;
  website: string | null;
  email: string | null;
  akreditasi: string | null;
  logo_sekolah_url: string | null;
  logo_dinas_url: string | null;
  nama_kepala_sekolah: string | null;
  nip_kepala_sekolah: string | null;
  ttd_kepala_sekolah_url: string | null;
  cap_sekolah_url: string | null;
  template_kartu_pelajar_depan_url: string | null;
  template_kartu_pelajar_belakang_url: string | null;
}